# Conflicts:
#	app/build.gradle
#	app/src/main/assets/nhentai.js
#	app/src/main/java/me/devsaki/hentoid/activities/sources/MultpornActivity.kt
#	app/src/main/java/me/devsaki/hentoid/activities/sources/NhentaiActivity.kt
#	app/src/main/java/me/devsaki/hentoid/activities/sources/YifferActivity.kt
#	app/src/main/java/me/devsaki/hentoid/database/DatabaseMaintenance.kt
#	app/src/main/java/me/devsaki/hentoid/enums/Site.kt
#	app/src/main/java/me/devsaki/hentoid/json/sources/deviantart/DeviantArtDeviation.kt
#	app/src/main/java/me/devsaki/hentoid/json/sources/ehentai/EHentaiGalleriesMetadata.kt
#	app/src/main/java/me/devsaki/hentoid/json/sources/ehentai/EHentaiImageResponse.kt
#	app/src/main/java/me/devsaki/hentoid/json/sources/kemono/KemonoGallery.kt
#	app/src/main/java/me/devsaki/hentoid/json/sources/nhentai/NHentaiContentMetadata.kt
#	app/src/main/java/me/devsaki/hentoid/json/sources/yiffer/YifferLoaderData.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/AllPornComicContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/DoujinsContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/EdoujinContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/EromangaContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/HdPornComicsContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/Hentai2ReadContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/HentaifoxContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/HiperdexContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/ManhwaContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/MultpornContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/MusesContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/NhentaiContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/NovelcrowContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/PixivContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/PorncomixContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/TmoContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/ToonilyContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/content/TsuminoContent.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/DeviantArtParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/DoujinsParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/EHentaiParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/EdoujinParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/EromangaParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/ExHentaiParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/HdPornComicsParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/Hentai2ReadParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/HentaifoxParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/HitomiParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/MangagoParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/Manhwa18Parser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/ManhwaParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/MrmParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/PixivParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/PururinParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/SimplyParser.kt
#	app/src/main/java/me/devsaki/hentoid/parsers/images/YifferParser.kt
#	app/src/main/java/me/devsaki/hentoid/util/AdBlocker.kt
#	app/src/main/java/me/devsaki/hentoid/workers/ContentDownloadWorker.kt
#	app/src/main/java/me/devsaki/hentoid/workers/TransformWorker.kt
#	app/src/main/res/raw/sites.json
#	app/src/main/res/values-es/strings_settings.xml
#	app/src/main/res/values-sv/strings_settings.xml
#	app/src/main/res/values/strings_settings.xml
#	app/update.json
#	gradle/wrapper/gradle-wrapper.properties

+++ app/src/main/java/me/devsaki/hentoid/util/file/ArchiveStreamer.kt
@@ -21,6 +21,7 @@ class ArchiveStreamer(
     context: Context,
     val archiveUri: Uri,
     append: Boolean,
+    val removeArchivedFiles : Boolean,
     val onProgressChange: ((Float) -> Unit)? = null
 ) {
 
@@ -95,7 +96,7 @@ class ArchiveStreamer(
                             }
                         } ?: throw IOException("Document not found : $uri")
                         Timber.d("Processing archive queue END : $uri")
-                        removeFile(context, uri)
+                        if (removeArchivedFiles) removeFile(context, uri)
                         // Only remove from queue if all above has succeeded
                         filesQueue.remove(uri)
